requirements.txt
dnspython
python-whois
jsonschema
datetime
requests
googlesearch-python
